refactor: keep Card document parser raw
