+++ src/entities/card/@x/study-progress.ts
@@ -1,2 +1,1 @@
-export type { CardDocument } from "../api/document";
 export type { CardId } from "../model/types";
