+++ src/entities/card/api/document.ts
@@ -1,8 +1,6 @@
 import { z } from "zod";
 
-import { mapStudyProgressDocument, type StudyProgress } from "@/entities/study-progress/@x/card";
 import { firestoreTimestampDateSchema, parseFirestoreDocument } from "@/shared/api";
-import type { CardId, RemoteCardRead } from "../model/types";
 
 const cardDocumentSchema = z.object({
   id: z.string().optional(),
@@ -26,37 +24,6 @@ const cardDocumentSchema = z.object({
 });
 
 export type CardDocument = z.infer<typeof cardDocumentSchema>;
-export interface CardReadModels {
-  card: RemoteCardRead;
-  progress: StudyProgress;
-}
 
-const parseCardDocument = (id: string, value: unknown): CardDocument =>
+export const parseCardDocument = (id: string, value: unknown): CardDocument =>
   parseFirestoreDocument(cardDocumentSchema, "card", id, value);
-
-const mapCardDocument = (id: CardId, document: CardDocument): RemoteCardRead => {
-  const card: RemoteCardRead = {
-    id,
-    frontText: document.frontText,
-    backText: document.backText,
-    tags: document.tags,
-    uniqueKey: document.uniqueKey,
-    deckId: document.deckId,
-    uid: document.uid,
-    createdAt: document.createdAt,
-    updatedAt: document.updatedAt,
-    deletedAt: document.deletedAt,
-  };
-  if (document.url !== undefined) card.url = document.url;
-  if (document.startLine !== undefined) card.startLine = document.startLine;
-  if (document.endLine !== undefined) card.endLine = document.endLine;
-  return card;
-};
-
-export const readCardDocument = (id: CardId, value: unknown): CardReadModels => {
-  const document = parseCardDocument(id, value);
-  return {
-    card: mapCardDocument(id, document),
-    progress: mapStudyProgressDocument(id, document),
-  };
-};
