+++ src/entities/card/model/dto.ts
@@ -1,0 +1,20 @@
+import type { CardDocumentFields, CardId, RemoteCardRead } from "./types";
+
+export const mapCardDocument = (id: CardId, document: CardDocumentFields): RemoteCardRead => {
+  const card: RemoteCardRead = {
+    id,
+    frontText: document.frontText,
+    backText: document.backText,
+    tags: document.tags,
+    uniqueKey: document.uniqueKey,
+    deckId: document.deckId,
+    uid: document.uid,
+    createdAt: document.createdAt,
+    updatedAt: document.updatedAt,
+    deletedAt: document.deletedAt,
+  };
+  if (document.url !== undefined) card.url = document.url;
+  if (document.startLine !== undefined) card.startLine = document.startLine;
+  if (document.endLine !== undefined) card.endLine = document.endLine;
+  return card;
+};
