+++ src/entities/card/model/types.ts
@@ -15,6 +15,7 @@ import type {
 export type RemoteCard = z.infer<typeof cardSchema>;
 type StudyProgressField = "score" | "numberOfSeen" | "lastSeenAt" | "nextSeeingAt" | "interval";
 export type RemoteCardRead = Omit<RemoteCard, StudyProgressField>;
+export type CardDocumentFields = Omit<RemoteCardRead, "id">;
 export type LocalCard = z.infer<typeof localCardSchema>;
 export type Card = RemoteCard | LocalCard;
 export type CardCreate = z.infer<typeof cardCreateSchema>;
